Reorganized MobileKeyboard component
